Change Turn to handle alternating teams, score limit.
GameManager now has a concept of the "buzzed team" which
should NOT be awarded with points. This is set as the
currently active team when the time runs out. Teams
now have a concept of a round score, since one time
through the Turn activity could be any number of turns
per team. Therefore, a round score was required for
each team.