Fixed some bugs with Turn. Added Fullscreen Skip button in FreePlay.
Reworked some copied code in Turn (showCard copied into doBack().
Moved card/deck logic out of Turn and into GameManager to help
with this change.

Handled some Bugs in onBack - when skipping, it returned to the
wrong team. It also got NEXT team instead of PREVIOUS... which is
fine in a 2 player game but breaks in 3+.