A lot of work to convert us to using json and organize our parsing.
I tried to separate all of the parsing logic into the PackParser and
the networking logic into the PackClient. There was necessary code to
be added in Deck to support this change (that is not done). I also
added CardJSONIterator to pull cards in in an efficient manner. I also
wrote a small python script for converting the xml into json and
removed the legacy JSON files.