Removed music and added turn ticking first pass.
There are a few known issues with this method that
I will try to work out on next commit. This check in
removed the music from the entire app. It removed the
preference for it, removed all overrides to onResume and
onPause to handle pausing music.

To handle ticking, I added a "state machine" class
that handles the ticking sounds as the game timer advances.